playlist-downloader/pyproject.toml
[project]
name = "playlist-downloader"
version = "0.1.0"
description = "CLI que baixa músicas de uma playlist (YAML) como MP3 via yt-dlp"
requires-python = ">=3.11"
dependencies = [
    "mutagen>=1.47",
    "pyyaml>=6.0",
    "yt-dlp>=2026.3.17",
]

[project.scripts]
playlist-downloader = "playlist_downloader.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
